Extract morning/evening AI briefings from App.jsx into useDailyBriefings hook
Moves generateMorningSummary, generateEveningReflection, their dismiss
callbacks, and both day-rollover reset effects out of App.jsx verbatim
into src/hooks/useDailyBriefings.js. The two briefings are structural
twins (data gathering, aiComplete call, per-day localStorage cache,
dismissal, rollover reset), so they extract as one cohesive hook.

Glance state stays owned by useVoiceAI and is passed in; consumers keep
receiving the callbacks through the context object unchanged. CLAUDE.md
called this candidate useMorningSummary — named useDailyBriefings since
it covers the evening reflection too.

Part of the App.jsx decomposition tracked in CLAUDE.md.
